Update logstash manifest for staging
If ETR is to send it's logs to elastic we need to expose logstash
outside of the cluster. logstash uses a TCP-based protocol and thus
we need to cater for that and expose the logstash TCP port using
service type LoadBalancer and specify a "static" IP rather than usa an
ingress with a host name (as done for http(s)).

Change-Id: I093af6af88cf231beaa024ee40b7d036fecc7078
Solves: Jira TEST-181