Update to the latest version of ETR on staging
This override is not the proper way of deploying to staging, however
the changes in the ETR are important and the current changes that
we have on github in the environment provider are too dangerous to speed
through without a long testing loop.
We want to get these ETR changes to production before that.

Change-Id: I982f24f4f47a076d7b1a61b112f6d2b8690267d2