Remove all patches for the environment provider for dev
Moved the old base and autoscaling components to their own directories
that shall be removed later. It was easier and less intrusive than
moving the patches around and trying to get them to work exactly as
before.
These oldbase and oldautoscaling components shall be removed when the
new release without a hosted environment provider has been created.

Change-Id: I030b41dfd9d7abba3c257d0d4e5bcbccbf2ee1de