Attempt to fix connection errors for ETOS
We are seeing a few connection errors when triggering an ETOS test-run.
I noticed a few restarts of the ETOS API with exit code 137 (OOMKill)
and I noticed that a pod was getting dangerously close to the memory
limit.
I increase the memory limit to see if that mitigates the problems we are
seeing. I also lowered the threshold for autoscaling based on memory and
I increase the minimum replicas for the ETOS API.

Change-Id: I8702dfa8ff687070412d7419000af5ad6d3d87fe