Attempt to fix the open telemetry operator
The open telemetry operator tries to create or communicate with
PodMonitor which is an object that we cannot create in our cluster.
It is unclear where this shall be fixed, but this update will tell the
collector to not want metrics with the hope that this tells the operator
to not care about PodMonitor objects.

Change-Id: Ib1733aa625554cda12aae6ac7c6d68222d76a3f1