Increase disk size for ETCD in production
We had a disk size of 1g, which can be considered quite low. Especially
if we are to use the database much more in the future.
I increased the default size to 10g on github, but I believe that we may
need even more than that for production.
According to the ETCD FAQ CoreOS uses a dedicated dual-core with 2g RAM
and 80g SSD at the very least. We are not there with the machine and I
don't think we have fast storage, but we should at least increase the
disk usage to match.
https://etcd.io/docs/v3.5/faq/

Change-Id: I5e02adb3ba1d6effc61d45943e0237327ca69265