Initial implementation of a testrun
This implements the controller for testrun, environment and providers.
Environment is only used for checking that providers are available and
the provider implementation will only check if external providers
respond properly on a healthcheck endpoint.

The testrun controller can run a suite runner if an environment reports
that it is ready and it will wait for the suite runner to complete and
delete it and environment after execution.
The testrun controller is quite hard-coded though and does not deploy
the log listener sidecar or any logstash sidecars for the suite runner.