Implement a cluster resource
The cluster resource will create all the services needed for an
ETOS deployment, including extra infrastructure needs if needed.
It is designed to make it a lot easier to get an ETOS cluster up
and running, especially when running ETOS in a local kubernetes
cluster.
Note that this cluster resource is not complete yet. It does not,
for example, deploy any providers for an ETOS testrun.
It does not read secrets that can be added in the cluster custom
resource.
It does not deploy a logstash instance for ETOS logging.