A working state of the testrun controller (#246)
This implements the controller for testrun, environment and providers.
Environment is only used for checking that providers are available and
the provider implementation will only check if external providers
respond properly on a healthcheck endpoint.

The testrun controller can run a suite runner if an environment reports
that it is ready and it will wait for the suite runner to complete and
delete it and environment after execution.
The testrun controller is quite hard-coded though and does not deploy
the log listener sidecar or any logstash sidecars for the suite runner.

The cluster resource will create all the services needed for an
ETOS deployment, including extra infrastructure needs if needed.
It is designed to make it a lot easier to get an ETOS cluster up
and running, especially when running ETOS in a local kubernetes
cluster.
Note that this cluster resource is not complete yet. It does not,
for example, deploy any providers for an ETOS testrun.
It does not read secrets that can be added in the cluster custom
resource.
It does not deploy a logstash instance for ETOS logging.