Restart deployments and statefulsets if configurations have changed (#398)
* Restart deployments and statefulsets if configurations have changed

By adding or updating an annotation to the spec Kubernetes
will automatically recreate the deployments and statefulsets
causing the pods to restart and get the new configurations.
This way of doing it is the same way as kubectl rollout restart.

* Set the annotation on the pod instead

Setting the annotation on the deployment/statefulset did not
automatically re-deploy the pods. Setting it on the podspec
does.