Enable opentelemetry support for ETOS when deployed with Cluster spec
Add a new parameter to the cluster spec called openTelemetry where it
is possible to pass in an enpoint and whether the endpoint is insecure
or not. This is how we use opentelemetry so this is not a complete
solution, but something to start with.

ETOS does not deploy an opentelemetry collector, only passes the
environment variables that the services use to connect and we expect the
system engineers to deploy the collectors separately.

I also removed the extra configuration that we could pass to the suite
starter and suite runner since they are now redundant.