Pass opentelemetry environment variables to v1alpha ETOS
The Suite Runner job will get its environment variables automatically
from the deployed ETOS cluster whereas for the environment requests we
check if we have a cluster to fetch and get it from there when possible.

I also make sure that we pass in the traceparents that can be set in the
annotations field of our testruns and environmentrequests so that the
opentelemetry trace graph is complete.

This change requires an update of the ETOS API as well as the
environment provider to be considered complete.