Upgrade kubebuilder v4.1.1 to v4.5.2
We've not done this in a while so there's a lot of changes made.
Basically I used kubebuilder alpha generate to create all new files and
structure that are created using the newer version. Then I diffed all
files with the ones we already had and merged them together.
I have not noticed any problems, everything builds OK and I was able to
start several testruns using the controller so I am quite certain this
works.

Note that most changes are automatic and I did not make any functional
changes from our side (except for a few golangci-lint complaints).
The newer version of kubebuilder also enabled a few github workflows
for testing so I fixed a few of the tests for us so they actually work.
I will be sending another PR where I increase the coverage of our tests
as well.