Fix the e2e tests so that they can be run multiple times
There were a few dangling resources causing us to have to
recreate the kind cluster every time we wanted to re-run tests.
These tests can now be run multiple times in a row.

I also enabled Prometheus in the e2e tests. This is not something we
need right now, but we will probably need it in the future and it does
not cost that much to add it.