Add e2e tests for the ETOS controller
The test will deploy the controller and a fully functioning
ETOS cluster and then deploy an ETOS testrun and wait for
it to succeed.
We only test the v1alpha version of ETOS at the moment but
we could extend the actual test that we run in the testrun
to start up a v0 testrun as well to verify that it works
as expected.