Upgrade kubebuilder v4.1.1 to v4.5.2 (#405)
* Upgrade kubebuilder v4.1.1 to v4.5.2

We've not done this in a while so there's a lot of changes made.
Basically I used kubebuilder alpha generate to create all new files and
structure that are created using the newer version. Then I diffed all
files with the ones we already had and merged them together.
I have not noticed any problems, everything builds OK and I was able to
start several testruns using the controller so I am quite certain this
works.

Note that most changes are automatic and I did not make any functional
changes from our side (except for a few golangci-lint complaints).
The newer version of kubebuilder also enabled a few github workflows
for testing so I fixed a few of the tests for us so they actually work.
I will be sending another PR where I increase the coverage of our tests
as well.

* Fix the e2e tests so that they can be run multiple times

There were a few dangling resources causing us to have to
recreate the kind cluster every time we wanted to re-run tests.
These tests can now be run multiple times in a row.

I also enabled Prometheus in the e2e tests. This is not something we
need right now, but we will probably need it in the future and it does
not cost that much to add it.

* Add conversion webhooks

This is not something we are using at the moment, but sadly with the old
version of kubebuilder these conversion webhooks were created
automatically when creating webhooks in general. That means that we have
a conversion webhook in the CRD spec and we are not allowed to update
the CRDs without recreating them (when running with apply
--server-side).

* Enable prometheus

* Remove a label for selector since it is immutable

* Reduce the length of a line to make make lint pass

* Fix the testrun_webhook so that it actually defaults to cluster

Added more logging as well and fixed a few spelling errors