Create a simple program for deploying ETOS to Kind for local testing
Can be deployed using make deploy-local and undeployed using
undeploy-local
Deploying a local ETOS this way will enable to use of ETOS locally but
will not work with the etos client at the moment since we do not deploy
an ingress.
It is still possible to create TestRun resources directly to the
cluster which allows us to test ETOS locally.