Refactor the environment provider controller
Added some more information fields on kubectl get
Added providers field to the environment resource so that we may check
provider status in the future. It is not yet populated.
Added a new Pending status for environments.
Fixed a nasty reconcile loop that could happen if the release job failed